refactor(config): split config.go into focused domain files
Mirrors the test file structure already in place (config_arch_test.go,
config_game_test.go, config_load_test.go, config_defaults_test.go,
config_clone_test.go). No public API or behavior changes.

- config.go: Load, handleReadError, migrateLyraKey, resolveEnginePath
- config_types.go: all struct definitions (Config through DDCConfig)
- config_game.go: GameConfig methods + ResolveServerBuildDir
- config_arch.go: NormalizeArch, ServerPlatformDir, BinariesPlatformDir, UEPlatformName
- config_defaults.go: Defaults()
- config_clone.go: Config.Clone()
